requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
PyMuPDF>=1.24.0
python-pptx>=0.6.23
python-docx>=1.1.0
Pillow>=11.0.0